scripts/requirements.txt
# Python dependencies for scripts
Pillow>=10.0.0
# Optional: exifread is an alternative EXIF library
# exifread>=3.0.0
